XIVY-2942 refactor: extract field detail content components
