+++ packages/dataclass-editor/src/components/dataclass/detail/field/FieldProperties.tsx
@@ -1,0 +1,11 @@
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@axonivy/ui-components';
+import { FieldModifierCheckbox } from './FieldModifierCheckbox';
+
+export const FieldProperties = () => (
+  <Collapsible defaultOpen={false}>
+    <CollapsibleTrigger>Properties</CollapsibleTrigger>
+    <CollapsibleContent>
+      <FieldModifierCheckbox label='Persistent' modifier='PERSISTENT' />
+    </CollapsibleContent>
+  </Collapsible>
+);
